+++ bootstrap.php
@@ -19,6 +19,10 @@ if ( function_exists( 'add_action' ) ) {
 					'name'     => 'migration',
 					'label'    => __( 'Migration', 'wp-module-migration' ),
 					'callback' => function ( Container $container ) {
+						if ( ! defined( 'NFD_MIGRATION_MODULE_VERSION' ) ) {
+							define( 'NFD_MIGRATION_MODULE_VERSION', '1.2.3' );
+						}
+						
 						$brand = $container->plugin()->id;
 						if ( 'atomic' === getContext( 'platform' ) ) {
 							$brand = 'bh-cloud';
